Reorganização dos controllers e dos models
