+++ database.json
@@ -2,18 +2,18 @@
     "professores": [
         {
             "data_nascimento": "1980-05-15",
-            "disciplina": "MatemÃ¡tica",
-            "nome": "Carlos Silva",
+            "disciplina": "MatemÃƒÆ’Ã†â€™Ãƒâ€šÃ‚Â¡tica",
+            "nome": "Polly",
             "salario": 5000.0,
-            "id": "ae06e4be-a879-43fb-89a4-41dc685714df"
+            "id": "9b4c879c-64db-4f7b-975d-401434cbe42c"
         }
     ],
     "turmas": [
         {
-            "capacidade": 30,
-            "nome": "Turma A",
-            "turno": "Manhã",
-            "id": "a2365b39-2066-4b4c-8b4b-3d50fa7d3418"
+            "capacidade": 40,
+            "nome": "Oi",
+            "turno": "noite",
+            "id": "25962539-47ef-4dfe-ba69-aab93d9eb269"
         }
     ],
     "alunos": [
@@ -22,7 +22,7 @@
             "nome": "Ana Souza",
             "nota_primeiro_semestre": 8.5,
             "nota_segundo_semestre": 9.0,
-            "id": "ea972168-2a00-48fa-83e5-884e23969182"
+            "id": "7b1844e7-148b-4686-95c9-78744e9a5e9f"
         }
     ]
 }
